base work for adding new APIs, no changes to default behaviour intented
