+++ .env.example
@@ -6,6 +6,9 @@ KICKTIPP_NAME_OF_COMPETITION=competition-name
 # optional environment variables
 KICKTIPP_HOURS_UNTIL_GAME=8
 KICKTIPP_RUN_EVERY_X_MINUTES=20
+OVERWRITE_TIPS=true # default=false
+ODDS_PROVIDER=kicktipp.de  # 'kicktipp.de' (default) oder 'the-odds-api.com'
+THE_ODDS_API_KEY=  # Nur erforderlich, wenn ODDS_PROVIDER='the-odds-api.com'
 
 ZAPIER_URL=https://hooks.zapier.com/hooks/catch/12345678/abcdefgh/
