+++ src/kicktipp_bot/core/quote_extractor_oddsapi.py
@@ -1,0 +1,12 @@
+import logging
+from typing import Optional
+
+logger = logging.getLogger(__name__)
+
+class QuoteExtractorOddsApi:
+    @staticmethod
+    def extract_quotes(*args, **kwargs) -> Optional[list]:
+        """Stub for extracting quotes from the-odds-api.com (to be implemented)."""
+        # TODO: Implement API call and extraction logic
+        logger.warning("Quote extraction from the-odds-api.com not yet implemented.")
+        return None
